Extract waitForDeployCreation function
