+++ src/wait-for-deploy-creation/index.js
@@ -1,0 +1,1 @@
+export { default } from './wait-for-deploy-creation'
